+++ buildbot.py
@@ -4,15 +4,16 @@
 import os
 import sys
 import json
+import shutil
 import subprocess
 
 project_name = 'recycle'
 
 
-def run_command(args):
+def run_command(args, env_ext={}):
     print("Running: {}".format(args))
     sys.stdout.flush()
-    subprocess.check_call(args)
+    subprocess.check_call(args, env=dict(os.environ.copy(), **env_ext))
 
 
 def get_tool_options(properties):
@@ -81,6 +82,19 @@ def install(properties):
     run_command(command)
 
 
+def cmake(properties):
+    build_path = 'build'
+    if os.path.exists(build_path):
+        print("Path '{}' already exists - removing".format(build_path))
+        shutil.rmtree(build_path)
+    os.mkdir(build_path)
+    old_cwd = os.getcwd()
+    os.chdir(build_path)
+    run_command(['cmake', '../'], env_ext={'VERBOSE': '1'})
+    run_command(['cmake', '--build', '.'], env_ext={'VERBOSE': '1'})
+    os.chdir(old_cwd)
+
+
 def coverage_settings(options):
     options['required_line_coverage'] = 100.0
 
@@ -103,6 +117,8 @@ def main():
         run_tests(properties)
     elif cmd == 'install':
         install(properties)
+    elif cmd == 'cmake':
+        cmake(properties)
     else:
         print("Unknown command: {}".format(cmd))